+++ HPL2/core/sources/graphics/MaterialType_Decal.cpp
@@ -96,6 +96,8 @@ namespace hpl {
 		cParserVarContainer defaultVars;
 		defaultVars.Add("UseUv");
 		defaultVars.Add("UseColor");
+
+		if(cGraphics::GetLinearSpaceRendering())	defaultVars.Add("LinearSpace");
 		
 		mpProgramManager->SetupGenerateProgramData(	eMaterialRenderMode_Diffuse,"Diffuse","deferred_base_vtx.glsl", "deferred_decal_frag.glsl", 
 											vDiffuseFeatureVec,kDiffuseFeatureNum, defaultVars);